feat: Replace float with uint64_t for improved precision
- Refactored calculations to use `uint64_t` instead of `float` and `double` for better accuracy.
- Updated struct definitions and function signatures across multiple files.
- Added utility functions for fixed-point arithmetic to support the new data type.
- Adjusted test cases to reflect changes in data representation.
- NOTE: `UTXOTxParameters` is not yet refactored and requires further updates.

+++ example/account_handling/AccountHelper.hpp
@@ -37,8 +37,8 @@ extern "C"
 #endif
     typedef struct DevConfig
     {
-        char  Addr[255];
-        float Cut;
+        char     Addr[255];
+        uint64_t Cut;
     } DevConfig_st2;
 
     typedef char AccountKey2[255];
@@ -61,7 +61,6 @@ namespace sgns
         }
 
         std::shared_ptr<TransactionManager> GetManager();
-
 
     private:
         std::shared_ptr<GeniusAccount>                             account_;
