+++ src/account/MintTransaction.cpp
@@ -9,7 +9,10 @@
 
 namespace sgns
 {
-    MintTransaction::MintTransaction( double new_amount, std::string chainid, std::string tokenid, const SGTransaction::DAGStruct &dag ) :
+    MintTransaction::MintTransaction( uint64_t                        new_amount,
+                                      std::string                     chainid,
+                                      std::string                     tokenid,
+                                      const SGTransaction::DAGStruct &dag ) :
         IGeniusTransactions( "mint", SetDAGWithType( dag, "mint" ) ), //
         amount( new_amount ),                                         //
         chain_id( chainid ),                                          //
@@ -41,15 +44,18 @@ namespace sgns
         {
             std::cerr << "Failed to parse TransferTx from array." << std::endl;
         }
-        double v64 = tx_struct.amount();
+        uint64_t    v64     = tx_struct.amount();
         std::string chainid = tx_struct.chain_id();
         std::string tokenid = tx_struct.token_id();
         //std::memcpy( &v64, &( *data.begin() ), sizeof( v64 ) );
 
-        return std::make_shared<MintTransaction>( v64, chainid, tokenid, tx_struct.dag_struct() ); // Return new instance
+        return std::make_shared<MintTransaction>( v64,
+                                                  chainid,
+                                                  tokenid,
+                                                  tx_struct.dag_struct() ); // Return new instance
     }
 
-    double MintTransaction::GetAmount() const
+    uint64_t MintTransaction::GetAmount() const
     {
         return amount;
     }
